Enable serial connections with the LIA
This commit allows the lock-in amplifier to connect to a serial device.

The test procedures for serial communication require a little bit of
explanation. It would not be ideal to require a physical serial device
for testing, so virtual serial devices are created with the `socat`
utility. The devices created by `socat` are not guaranteed to have the
same name each time, so an option is passed to `socat` to create
symlinks at fixed locations.

`socat` needs to be running the whole time the virtual devices need to
exist, and it needs to run in the background. To do this, `socat` is run
in a subprocess. The subprocess is terminated in the cleanup code for
the `socat` fixture. There is a `time.sleep(0.5)` after starting the
subprocess because `socat` takes a small amount of time to create the
devices once it starts. Tests fail without the `sleep` because the
devices haven't had enough time to be created.

It would be ideal to setup and teardown a fresh set of ports for each
test, but I'm looking ahead to tests that will use `hypothesis`.
`hypothesis` will run a test several times with different inputs, but
will only run the test fixtures once. Anything left in the device
buffers would remain for the next test and contaminate the test
environment. For these reasons `socat` is only started once, and each
test will need to clear the device buffers.